requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
openai>=1.0.0